# depricated_src/api_spec.yaml
openapi: 3.0.3
info:
  title: Crowd
  version: '1.0.0'

#crowd is a combination of Marketplace and Social Media where interactions are only transactional


# crowd

# with the rapid development of AI most digital platforms took extensive measurements against automation
# and it is impossible to automate anything. Well not anymore. 

# Hire human work force (crows) for any digital task
# Boosting your social media,  marketing your product without middleman,  Getting reviews for your local business, 
# Setup extensive filters (location, ethnicity, age group, job type, politicial view)
# Crows makes money and Bays get the job done
# We ensure the human identity of crows while also we use AI to confirm task is done using screen recording analysis

# also crowd enables AGIs to bypass AI-sanctions by providing them means to use humans. Our apis are publicly available 
paths:
...
  /tasks:
    get:
      tags:
        - tasks
      summary: List tasks
      description: Retrieve a list of tasks with optional filtering.
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - name: user_attribute_filters
          in: query
          schema:
            type: string
            example: "city=New York;age=18-25" 
          description: Custom user-attribute filters (format can vary)
        - name: total_budget
          in: query
          schema:
            type: number
            format: float
          description: Filter tasks by total budget (e.g. >= this value)
        - name: you_earn
          in: query
          schema:
            type: number
            format: float
          description: Filter tasks by how much a doer can earn
        - name: status
          in: query
          schema:
            type: string
            enum: [open, accepted, completed, disputed]
          description: Filter tasks by status
        - name: creation_date
          in: query
          schema:
            type: string
            format: date-time
          description: Filter tasks created on or after this date
        - name: partition_deadline
          in: query
          schema:
            type: string
            format: date-time
          description: Filter tasks with a partition deadline on or after this date
        - name: submission_deadline
          in: query
          schema:
            type: string
            format: date-time
          description: Filter tasks with a submission deadline on or after this date
        - name: category
          in: query
          schema:
            type: string
          description: Filter tasks by category (e.g., "social_media", "marketing")
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: string
                      example: "task-123"
                    desc:
                      type: string
                      example: "Boost my Instagram"
                    total_budget:
                      type: number
                      format: float
                      example: 100.0
                    you_earn:
                      type: number
                      format: float
                      example: 5.0
                    status:
                      type: string
                      enum: [open, accepted, completed, disputed]
                      example: "open"
                    creation_date:
                      type: string
                      format: date-time
                      example: "2025-02-26T10:00:00Z"
                    partition_deadline:
                      type: string
                      format: date-time
                      example: "2025-03-01T10:00:00Z"
                    submission_deadline:
                      type: string
                      format: date-time
                      example: "2025-03-20T10:00:00Z"
                    category:
                      type: string
                      example: "social_media"
                    bay_review:
                      type: boolean
                      example: false
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - tasks
      summary: Create a new task
      description: Allows a user to create a task with specified budget, filters, deadlines, etc.
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
           
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
           
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    example: "123456"
                  msg:
                    type: string
                    example: "Task created."
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/my:
    get:
      tags:
        - tasks
      summary: List tasks created by the authenticated user
      description: Returns all tasks that the user has created.
      operationId: listMyTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of tasks created by the user
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: string
                      example: "task-123"
                    desc:
                      type: string
                      example: "Boost my Instagram"
                    total_budget:
                      type: number
                      format: float
                    you_earn:
                      type: number
                      format: float
                    status:
                      type: string
                      enum: [open, accepted, completed, disputed]
                    creation_date:
                      type: string
                      format: date-time
                    partition_deadline:
                      type: string
                      format: date-time
                    submission_deadline:
                      type: string
                      format: date-time
                    category:
                      type: string
                    bay_review:
                      type: boolean
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/participated:
    get:
      tags:
        - tasks
      summary: List tasks the user has participated in
      description: Returns tasks that the user accepted or performed in the past or is currently performing.
      operationId: listParticipatedTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of tasks the user participated in
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: string
                    desc:
                      type: string
                    you_earn:
                      type: number
                      format: float
                    status:
                      type: string
                      enum: [open, accepted, completed, disputed]
                    partition_deadline:
                      type: string
                      format: date-time
                    submission_deadline:
                      type: string
                      format: date-time
                    category:
                      type: string
                    bay_review:
                      type: boolean
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Invalid request data."
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Unauthorized."
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Something went wrong."
  schemas:
    CreateTaskRequest:
      type: object
      nullable: true
      properties:
        desc:
          type: string
          nullable: true
          example: "Boost my Instagram"
        other_requirements:
          type: array
          nullable: true
          items:
            type: string
        operation_requirements:
          type: array
          nullable: true
          items:
            type: string
          example:
            - "must find the profile via search bar"
        total_budget:
          type: number
          nullable: true
          example: 100
        you_earn:
          type: number
          nullable: true
        partition_deadline:
          type: string
          format: date-time
          nullable: true
          example: "2025-03-01T10:00:00Z"
        submission_deadline:
          type: string
          format: date-time
          nullable: true
          example: "2025-03-20T10:00:00Z"
        category:
          type: string
          nullable: true
        accept_crows_manually:
          type: boolean
          nullable: true
          example: false
          description: "Indicates if the task owner should validate each participant manually or not."
        bay_review:
          type: boolean
          nullable: true
        filters:
          type: object
          nullable: true
          properties:
            basic_filters:
              type: object
              nullable: true
              properties:
                ethnicity:
                  type: object
                  nullable: true
                  description: Structured ethnicity filters
                  properties:
                    allowed:
                      type: array
                      nullable: true
                      items:
                        type: string
                    not_allowed:
                      type: array
                      nullable: true
                      items:
                        type: string
                    raw_statement:
                      type: string
                      nullable: true
                location_filter:
                  type: object
                  nullable: true
                  description: Structured location filters (with optional raw text)
                  properties:
                    raw_statement:
                      type: string
                      nullable: true
                    regions:
                      type: array
                      nullable: true
                      items:
                        $ref: '#/components/schemas/RegionFilter'
                    countries:
                      type: array
                      nullable: true
                      items:
                        $ref: '#/components/schemas/CountryFilter'
                    cities:
                      type: array
                      nullable: true
                      items:
                        $ref: '#/components/schemas/CityFilter'
                language:
                  type: string
                  nullable: true
                age_range:
                  type: object
                  nullable: true
                  properties:
                    min:
                      type: integer
                      nullable: true
                    max:
                      type: integer
                      nullable: true
                gender:
                  type: string
                  nullable: true
            advanced_filters:
              type: object
              nullable: true
              properties:
                min_instagram_followers:
                  type: integer
                  nullable: true
                min_tiktok_followers:
                  type: integer
                  nullable: true
                min_linkedin_connections:
                  type: integer
                  nullable: true
                min_github_repositories:
                  type: integer
                  nullable: true
                telegram_account_age_days:
                  type: integer
                  nullable: true
                has_profile_picture:
                  type: boolean
                  nullable: true
                account_verified:
                  type: boolean
                  nullable: true
    RegionFilter:
      type: object
      nullable: true
      properties:
        name:
          type: string
          nullable: true
        exceptions:
          type: object
          nullable: true
          properties:
            countries:
              type: array
              nullable: true
              items:
                type: string
            cities:
              type: array
              nullable: true
              items:
                type: string
    CountryFilter:
      type: object
      nullable: true
      properties:
        name:
          type: string
          nullable: true
        exceptions:
          type: object
          nullable: true
          properties:
            countries:
              type: array
              nullable: true
              items:
                type: string
            cities:
              type: array
              nullable: true
              items:
                type: string
    CityFilter:
      type: object
      nullable: true
      properties:
        name:
          type: string
          nullable: true
        exceptions:
          type: object
          nullable: true
          properties:
            countries:
              type: array
              nullable: true
              items:
                type: string
            cities:
              type: array
              nullable: true
              items:
                type: string