# depricated_src/models/create_task_request.py
# coding: utf-8

"""
    Crowbay

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, StrictBool, StrictFloat, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List, Optional, Union
from models.create_task_request_filters import CreateTaskRequestFilters
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class CreateTaskRequest(BaseModel):
    """
    CreateTaskRequest
    """ # noqa: E501
    desc: StrictStr
    other_requirements: Optional[List[StrictStr]] = None
    operation_requirements: Optional[List[StrictStr]] = None
    total_budget: Union[StrictFloat, StrictInt]
    you_earn: Optional[Union[StrictFloat, StrictInt]] = Field(default=None, description="How much each participant will earn")
    partition_deadline: Optional[datetime] = None
    submission_deadline: Optional[datetime] = None
    category: Optional[StrictStr] = None
    accept_crows_manually: Optional[StrictBool] = Field(default=None, description="Indicates if the task owner should validate each participant manually or not.")
    bay_review: Optional[StrictBool] = Field(default=None, description="Indicates if the task requires a special review")
    filters: Optional[CreateTaskRequestFilters] = None
    __properties: ClassVar[List[str]] = ["desc", "other_requirements", "operation_requirements", "total_budget", "you_earn", "partition_deadline", "submission_deadline", "category", "accept_crows_manually", "bay_review", "filters"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of CreateTaskRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of filters
        if self.filters:
            _dict['filters'] = self.filters.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of CreateTaskRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "desc": obj.get("desc"),
            "other_requirements": obj.get("other_requirements"),
            "operation_requirements": obj.get("operation_requirements"),
            "total_budget": obj.get("total_budget"),
            "you_earn": obj.get("you_earn"),
            "partition_deadline": obj.get("partition_deadline"),
            "submission_deadline": obj.get("submission_deadline"),
            "category": obj.get("category"),
            "accept_crows_manually": obj.get("accept_crows_manually"),
            "bay_review": obj.get("bay_review"),
            "filters": CreateTaskRequestFilters.from_dict(obj.get("filters")) if obj.get("filters") is not None else None
        })
        return _obj